+++ core/sail/extensible-store/src/test/java/org/eclipse/rdf4j/sail/extensiblestoreimpl/TransactionIsolationAndWalTests.java
@@ -7,74 +7,16 @@
  *******************************************************************************/
 package org.eclipse.rdf4j.sail.extensiblestoreimpl;
 
-import org.eclipse.rdf4j.IsolationLevels;
-import org.eclipse.rdf4j.model.vocabulary.RDFS;
-import org.eclipse.rdf4j.repository.sail.SailRepository;
-import org.eclipse.rdf4j.repository.sail.SailRepositoryConnection;
-import org.junit.Ignore;
 import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import java.util.concurrent.atomic.AtomicBoolean;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
 
 public class TransactionIsolationAndWalTests {
 
 	private static final Logger logger = LoggerFactory.getLogger(TransactionIsolationAndWalTests.class);
 
-	/*
-	 * Checks that there is no leak between transactions. When one transactions adds a lot of data to the store another
-	 * transaction should see either nothing added or everything added. Nothing in between.
-	 */
 	@Test
-	public void testReadCommittedLargeTransaction() throws InterruptedException {
-		SailRepository repository = new SailRepository(new ExtensibleStoreImplForTests());
-
-		int count = 100000;
-
-		AtomicBoolean failure = new AtomicBoolean(false);
-
-		Runnable runnable = () -> {
-
-			try (SailRepositoryConnection connection = repository.getConnection()) {
-				while (true) {
-					connection.begin(IsolationLevels.READ_COMMITTED);
-					long size = connection.size();
-					connection.commit();
-					if (size != 0) {
-						if (size != count) {
-							logger.error("Size was " + size + ". Expected " + count);
-							failure.set(true);
-						}
-						break;
-					}
-					Thread.yield();
-				}
-			}
-		};
-
-		Thread thread = new Thread(runnable);
-		thread.start();
-
-		try (SailRepositoryConnection connection = repository.getConnection()) {
-			connection.begin(IsolationLevels.READ_COMMITTED);
-			for (int i = 0; i < count; i++) {
-				connection.add(RDFS.RESOURCE, RDFS.LABEL, connection.getValueFactory().createLiteral(i));
-			}
-			connection.commit();
-
-			assertEquals(count, connection.size());
-
-		}
-
-		thread.join();
-
-		assertFalse(failure.get());
-
-		repository.shutDown();
+	public void noTestsHereYet() {
 
 	}
